+++ src/main/kotlin/kr/pincoin/api/infra/user/repository/UserJpaRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserJpaRepository : JpaRepository<UserEntity, Long> {
+interface UserJpaRepository : JpaRepository<UserEntity, Int> {
     fun existsByEmail(email: String): Boolean
 }
